WL#14637: allow one or more certificate file pointers
Node.js TLS-enabled APIs usually follow a pattern where certificates are
provided using explicit direct calls to "fs.readFileSync()" to read PEM
content beforehand.

Up until now, when creating a database connection with
Connector/Node.js, one is able to provide a path to a certificate
authority file (hopefully containing the entire chain) and additionally,
a path to a certificate revocation list file. Due to limitations in how
Node.js interacts with OpenSSL, the latter only worked as expected when
the certificate authority was the root itself. The only way to ensure
a certificate revocation list from an intermediate certificate authority
is validated is by using a list of all the CRL files in the chain.

To allow this, the connection should be able to be created using a list
of read via "fs.readFileSync()", following the pattern described above.

This patch ensures that connections can be created using a configuration
object, where the certificate authority chain and certificate revocation
list chain can be specified using a list of file pointers, effectively
closing BUG#25589348.