tests: adapt tests to proper index matching prepared statements
The optimizer introduced some changes that are causing some prepared
statements using generated columns, such as JSON indexes, to result in a
type mismatch. This happens due to some limitations in comparing JSON
values (the ones extracted from document fields) and SQL values (the
ones used as the index) and is particularly an issue whilst comparing
opaque JSON types and VARBINARY SQL types.

An example of such an occurence is when a statement uses the document
"_id" property, which is automatically used to create a PRIMARY KEY
index once a collection is created.

There is a workaround for this kind of limitation which consists in
extracting the SQL value from a JSON value using "JSON_UNQUOTE()". The X
Plugin is now using this strategy to ensure there an effective match
with the index values in cases where that might have not been happening.

This patch updates some tests to ensure any assertion of server-side
generated statements accounts for the changes (the introduction of
"JSON_UNQUOTE()") and includes a new formal regression test for the most
obvious case that could eventually break existing applications.