BUG#33740190: indexes for upstream document Date instances
When adding a document to a collection, applications can use JavaScript
Date instances to represent JSON date string values. The client simply
calls the ".toJSON()" method which returns a UTC date string using the
Zulu Time indicator "Z", as specified by the ISO 8601 standard.

However, the Zulu Time indicator is not covered by the SQL standard
and thus is not allowed by the MySQL optimiser. This means that it is
not possible to create a collection index using a virtual column with
the DATETIME (or similar) data type. Trying to do so yields the
following error:

  Error: Incorrect datetime value

This patch addresses the issue by replacing the Zulu Time indicator by
an explicit displacement indicator of "+00:00" which has the exact same
meaning from the ISO 8601 standpoint and is also allowed by the MySQL
optimiser.