BUG#33761268: skip connection requests on unavailable pool
When a connection pool is full, any new connection request is retried
within a sliding window of time until it reaches the time specified by
the "queueTimeout" option. By default, the connection request will be
retried forever, until the connection is effectively acquired.

Currently, if the pool is closed in the meantime, all connection
requests that have been queued will still be retried in the future, as
long as the application process does not terminate. This is bad and
might result in dangling sockets and formal MySQL server connections,
which is a waste of resources and a potential source of memory leaks.

This patch introduces the changes necessary to ensure all queued
connection requests in a pool are skipped when the pool is closed or is
in the process of being closed (where it should still no longer be
available in order to avoid parallelism issues).