deps: remove timekeeper
The timekeeper package is being used to perform fake time travel in a
couple of unit tests. The same can also be easily achieved by using a
mock of the JavaScript API resorting to other 3rd-party libraries
already being used like testdouble.js.

This patch introduces some changes to centralize the most common use of
the JavaScript API, in the form of "Date.now()" to a custom "global"
utility function as part of the existing system tools. Additionally, it
changes the corresponding component unit tests to use a fake version of
that function which can be controlled individually for each test and
avoids the need to use the timekeeper package. As a consequence, the
package was also removed from the list of "devDependencies".