misc: fix copyright tooling issues
The existing copyright tooling did not account for new or modified files
yet to be committed, only those that have been previously committed.
This patch addresses this limitation and ensures that the copyright
header/notice is poperly verified on all files that have been staged for
commit in the local git index.

Additionally, the use of the "Array.prototype.flat()" method was
incompatible with older Node.js versions. Since the same behaviour can
be easily reproduced using "Array.prototype.reduce()", the implementation
was changed, in particular, to allow the test suite to still be
compatible with those older Node.js versions.

Change-Id: Idba970dcbefb7f6b7011891fecdc3ec19a99d258