BUG#33940584: trust server on missing placeholder values
Placeholder arguments can be assigned with the "bind()" method in CRUD
operations by individually setting the value of each placeholder or by
performing multiple assignments using an object where fields and values
accordingly correspond to placeholder names and values.

With plain SQL statements, ordinal placeholder values are assigned
incrementally by calling "bind()" multiple times or by providing an
array of values as its arguments.

Currently, if that contract is broken somehow, the client will in some
specific cases yield an unexpected error. In particular, when the
placeholder value is undefined.

This is confusing, and since the X Plugin already accounts for
inconsistencies in placeholders and their values due to the typed nature
of protocol buffers, we can simply trust it will handle those and yield
any potential error.

This patch updates the client placeholder assignment logic to ensure
that undefined placeholder values are ignored and externalizes any
validation task to the server, where the X Plugin will yield errors
accordingly. Those errors are then relayed to the application in a more
consistent feedback loop.

Change-Id: I942b4172cd05ffa8a6af658a2852f8a409241460