tests: remove failing Scalar-encoding unit test
With the previous fix for BUG#33940584, "undefined" JavaScript values
should no longer be encoded as a specific type of
Mysqlx.Datatypes.Scalar message. This was not the case before, and there
was a unit test to check if those values could be encoded or not by that
type of message. This unit test was neither changed nor removed and
started to fail after the corresponding patch was introduced.

This patch removes the failing test, which does not have a meaning
anymore.

Change-Id: Iddf348d107450c1c64ee0843cccb25efc0335a03