tests: fix IPv6 functional tests with IPv4 host
Functional tests for IPv6 connections use a wrappper over the core
Node.js "dns.lookup()" API. This utility operates by resolving a
hostname to its corresponding IPv6 address. By default, the test
harness injects a hostname based on the value of the MYSQLX_HOST
environment variable, or by omission "localhost".

Although "localhost" works fine, assigning an IPv4 address to the
MYSQLX_HOST is also a common practice, and "dns.lookup()" is incapable
to resolve an IPv4 address to its corresponding IPv6 address. This
causes the IPv6 connection tests to fail.

This patch addresses this issue and updates the test suite alongside
the IPv6 address resolution utility to be able to work with both
common hostnames and IPv4 addresses.

Change-Id: I1037e6e3914b9dadfa1d7ba3d8433117a602db8d