misc: remove .gitreview configuration file
The .gitreview file should not be under source control because it is
only meaningful in a local development environment. Since it was
also included in the .gitignore file, it was assumed that would not
happen, but it was not the case.

This patch removes the .gitreview file from source control and also
removes the corresponding useless pointer from the .gitignore file.

Change-Id: I046dbbdc0bf9cf9644c92b6d206d45e7332b8c70