BUG#34016587: fix precision loss for DECIMAL values
Values stored in numeric column data types are being retrieved either as
JavaScript strings or numbers depending on whether there is a chance of
losing precision or not. However, this behavior is not consistent in the
case of DECIMAL values where the decimal part (the digits to the right
of the decimal point) cannot be accommodated by a JavaScript number.

In this case, unless the integer part cannot also be accommodated, the
value is effectively retrieved as a JavaScript number and it looses
decimal precision.

This patch introduces the changes to address this issue ensuring DECIMAL
values are returned as a JavaScript string when there is a chance of
losing precision.

Change-Id: I9a494059d9d902b0169beda0abd1df5d20c0d080