deps: avoid peer dependency mismatch
The "@typescript-eslint/eslint-plugin" library is a peer dependency of
"eslint-config-standard-with-typescript", however, the latter specifies
a semver range that includes only 4.x.x versions.

Although it is usually not a problem, the fact is that, starting with
npm v7, "npm install" will fail by omission if peer dependencies are
not within the semver range specified by the host/dependent package.

https://blog.npmjs.org/post/626173315965468672/npm-v7-series-beta-release-and-semver-major

This patch updates the version of "@typescript-eslint/eslint-plugin" to
ensure it is within the semver range specified by the current version of
"eslint-config-standard-with-typescript".

Change-Id: I8032898d2fdba2baec3ce70c8179922735c32385