tests: skip IPv6 tests with IPv4 host
The work done as part of revision 15ed85eb was meant to prevent the IPv6
functional tests to fail when the MYSQLX_HOST environment variable
pointed to an IPv4 address (and not a common hostname). It introduced
some changes that resulted in over-complicated fixtures that supposedly
allowed to work around those test failures.

This was not really the case, because the "dns.lookup()" API returns
only one IPv6 address which, in the case of a more convoluted setup,
where multiple addresses can be available (even virtual ones) like the
case of a NAT, will result in authentication failures due to the fact
that the MySQL server is not able to perform the reverse lookup.

Additionally, the "dns.lookup()" works as an identity function when the
argument is an IPv4 address, which only leads to more confusion.

When using an IPv4 address as the MYSQLX_HOST value, there is really not
point in trying to run IPv6 tests at all. So, somehow "forcing" them to
run using their IPv4 counterpart is not a great idea.

This patch reverts some of the changes introduced in revision 15ed85eb
and simplifies the IPv6 test setup by ensuring that the record family of
the address returned by "dns.lookup()" is, in fact the one for IPv6
addresses. If not, the lookup will result in an "undefined" address
which will let the test harness naturally skip those tests.

Change-Id: I30099a1c10d10726ffc40989a4720cff4baab237