tests: fix protobuf String message toJSON() unit test
The Mysqlx.Datatypes.Scalar.String adapter API contains a "toJSON()"
method which returns a different value depending on the collation used
to encode the value. With a "utf8mb4" collation, the underlying value is
decoded as a JavaScript utf8 String, whereas with a "binary" collation,
the value is decoded a JavaScript base64 String.

The existing unit test was focused on the case of a "utf8mb4" collation
but was retrieving a random collation id from the internal collation
mapping. This means it risked using a "binary" collation id causing the
test to fail.

In this specific case, the method uses two additional API methods,
"getCollationId()" and "toString()" which are already being tested,
which means the test can simply inject a fake version of both these
methods.

This patch updates the test to use a fake version of the methods already
being tested and decouples the test logic from the implementation
details of each of those, at the same time, removing the need to rely on
a valid collation id to execute the test.

Change-Id: I10c823c211357ba9386e94bb16d21908ec587a75