tests: simplify dummy socket timeout servers
The dummy socket servers used in functional tests for checking the
connection timeout, in particular, for the multi-host scenario, have a
somewhat complicated implementation. The resulting setup is overwhelming
and seems to be causing some synchronicity issues in the test harness.

These issues are, at times, observable when checking if the elapsed time
whilst trying to connect to the server is, in fact, above or equal to
the sum of all the reported connection timeouts.

This patch simplifies the dummy server setup and inverts the
timestamp-based checks to compare the elapsed time during each test to
the timeout itself, instead of simply comparing raw timestamps, which
are not easy to understand when the assertions fail.

Change-Id: I96ed9a4ab06eca9545a6f1b2ce4a922c42892537