WL#15133, BUG#29178528: narrow expression parser responsability
The X DevAPI expression parser is performing two different duties when
it should only have a single responsability. In this case, besides
parsing an expression string, it also creates the corresponding protobuf
message instance using the available protocol stubs.

From an architectural standpoint, the parser should be an isolated
component and should not have to know about the existence of an
underlying protocol. Additionally the parser output should not be tied
to a format defined by a 3rd-party library.

This patch narrows the job of the expression parser component to the
single responsability of parsing expression strings and returning a
basic structured output and moves, to the protocol stub adapters, the
responsability of actually encoding the resulting protobuf message
instances. As a result, it also decouples the parser from the protocol
itself and from the custom restricted format specified by the
google-protobuf library.

As a consequence of the structural changes in the expression parser,
some of its known limitations have also been addressed. In particular,
the one described by BUG#29178528.

Change-Id: I1550ac72ee288864eaada69e6630d2441274b3a0