BUG#29795595: allow leading and trailing whitespaces in expressions
The X DevAPI expression parser is pretty strict when it comes to leading
and trailing whitespaces in an expression string.

These extra characters do not have any special meaning and do not cause
any particular harm.

This patch relaxes this behavior and updates the expression parser to
allow extra leading and trailing whitespaces in expression strings.

Change-Id: I74d279e5db323c420c8cb4ff42b5831842c1b4a1