BUG#27361584: prevent invalid usage of globstars
Globstars ("**") can be used to create an implicit link between two
elements of a document path. They make it easier and less verbose to
perform operations on a deeper level of a document. However, they are
only valid in a specific setting where the both the leading and trailing
elements in the document path are correctly specified, like the
following:

  '$.foo**.bar'
  '$**.bar'

In practice, this means they cannot be used as the last item in a
document path.

The expression parser does not account for invalid usage of a globstar,
for instance, like in the following instances.

  '$.foo.**bar'
  '$.foo**'
  '$**'

This patch addresses this issue and updates the sub-parser for
"documentPath" expressions to explicitely fail for the cases where a
globstar is the last item in such an expression.

Change-Id: If16c9e1d81af09e912026a566ecb4be173585c89