BUG#31017606: fix JSON_UNQUOTE shorthand syntax result
The "JSON_UNQUOTE" function returns a utf8mb4 string. Since "->>" is
simply a shorthand syntax calling "JSON_UNQUOTE" on top of the value of
a document field, it should also follow the same convention.

Contrary to "JSON_EXTRACT", where an expression can simply be
interpreted as the column identifier itself, an expression containing
the "->>" shorthand syntax must be encoded as a function call, where the
sole parameter is the corresponding column identifier / document field.

Right now, expressions using this shorthand syntax are not being
correctly interpreted or encoded.

This patch addresses this issue and ensures that the results of calling
the "JSON_UNQUOTE" function or using the "->>" corresponding shorthand
syntax are consistent.

Change-Id: I1a030249a9dddc88eaf0810fbe75f9eeb8d19afc