WL#12019: better handling of X DevAPI expressions
A single generic parser was being used to parse all the X DevAPI
expressions in the scope of a given statement right before it was
executed.

Although this centralized approach looks good on paper, it also brings
with it a lot of unwanted limitations and performance issues. With a
dynamicaly generated parser, this means that to overcome specific
diferences between parser types, the parser has to be re-created for
each different type that requires so. Also, not spreading the parsing
tasks between stages in the statement life-cycle means we are creating
a possible bottleneck in a single stage.

Additionally, the fact that the "mysqlx.expr()" method is not only
parsing but also creating the corresponding protobuf message of a given
expression means that we need to take special care to handle expressions
that are already in the protobuf format when we have to parse any other
expression in the scope of the statement execution.

The fact that expressions are also being parsed at the protocol "layer"
in a conceptual connector stack is also not good from a design
standpoint. This also leads to a situation where expressions are parsed
more than once when using prepared statements. X DevAPI expressions are
a high level concept that should be handled at the API level.

The expression parser itself is also stateful because it needs to keep
track of the placeholders for each expression. This should be a
responsability for the routines that create the protobuf messages.

This patch addresses these issues by narrowing the scope and
responsability of the "mysqlx.expr()" method, which now only works as an
abstraction for a given value in the form of an expression string and by
moving the parsing logic to the API level. It introduces some new
DevAPI-level abstractions that encapsulate their own parser setup, which
is created only once and used every time the application calls an API
method that relies on structured expressions. Internally, the parser
implementation is now stateless and the placeholder assignment logic has
been simplified and moved to the proper layer in the stack.

To ensure this changes are, in fact, meaningful, the patch also
introduces a few benchmarks for the relevant CRUD-like statements.

Change-Id: I37c63840cc00b3c006b0839a47d31ada250c0cab