WL#15434: update 3rd-party dependencies for 8.0.32
This patch updates all 3rd-party dependencies with new released versions
that are compatible with the current minimum required Node.js version
(v14.0.0).

As part of the "google-protobuf" update, the X Protocol stubs have been
re-generated using the corresponding version of the compiler.

The latest version of "dns2" also introduces a few breaking changes
which lead to a few small changes in the test suite.

Change-Id: Ia0dade4e2bb256ffebb7abdf6b67255257dffad7