WL#15353: add support for running tests in a container
In order to run the automated test suite, and besides a running MySQL
server instance, users need to have a previously installed compatible
Node.js engine version and, additionally, need to install the required
3rd-party dependencies.

Although very little, this process always requires a bit of Node.js
expertise, and can be a hurdle for developers that want to test specific
server features, and not the corresponding client/connector UX.

One way to introduce a small shortcut and reduce the burden of setting
up a local testing environment is by leveraging docker containers. In
this case, users would only need to install the container engine, and
any additional setup would be performed automatically and without
requiring any specific language/platform knowledge apart from knowing to
execute a shell script in the system and launching the MySQL server.

This patch introduces tooling that facilitates the task of running the
test suite using a docker container and a running MySQL server instance
on Unix-based environments (such as Linux and macOS). It delivers a
shell script that automates the task of building the container image and
run it using the exact same configuration parameters that can be
specified whilst in the local system environment.

At the same time, it removes some older tooling used to facilitate a
similar task which was not in any way tied to the existing public npm
scripts and was, at best, a source of confusion.

Change-Id: I2a5fdaf8edb15fd6be5f70d7457a293fe7d1d6e3