BUG#34728259: return unsafe numeric field values as strings
When retrieving records in a JSON column, if a field contains an unsafe
numeric value, it is parsed as a JavaScript number from the
corresponding result set. This means that applications are dealing with
values that are loosing precision.

The reason for this to happen is that the connector is parsing JSON
strings as their corresponding plain JavaScript object counterparts for
the sake of convenience. This creates a problem wherein the native
"JSON.parse()" API automatically coerces a numeric value of a field to
the a JavaScript number and there is no way to override that behavior,
even using a "reviver", because the value is already coerced by that
point.

This patch addresses the issue by introducing a 3rd-party JSON parser
which is capable of "reviving" unsafe big numbers (which loose precision
when represented by JavaScript numbers), or even all numeric values as
JavaScript strings. This is the strategy already in place to parse
unsafe numeric values from explicit column data types such as BIGINT and
DECIMAL.

Change-Id: I30ad08d6088cf4d7388ec78d71c63119de5e34ad