BUG#34767204: store lossless numeric field values
When a document contains numeric field values which are not safely
represented by a JavaScript number, it should still be possible to
save those values without loosing precision by using a textual JSON
representation of the document for adding it to a collection.

This was not the case, and all unsafe numeric values in a JSON document
were being truncated or rounded to the nearest value capable of being
represented by a JavaScript number. To make matters worse, there is no
guarantee that this value is itself safe for performing any kind of
mathematical operation.

The root cause that leads to this issue is the fact that JSON string
documents are being converted to the corresponding plain JavaScript
object by "collection.add()" via the native "JSON.parse()" method,
which "revives" all numeric values as JavaScript numbers.

This patch addresses this issue by replacing the native JavaScript JSON
parser with a bespoke version, based on a 3rd-party library, that is
capable of "reviving" unsafe numeric values as JavaScript strings.

Change-Id: I6c20e9cc83befcbaa2e5101ddb8be99e3a34d1ba