WL#15436: simplify result set row decoding
The Mysqlx.Resultset.Row message is not a structured protobuf message
message, but instead a binary field that is encoded using the primitive
protobuf "bytes" type.

In order to decode the content of each row in a result set, the client
needs to parse the content of this field using the additional metadata
available for a given column which is sent in the
Mysqlx.Resultset.ColumnMetadata message. This involves using the
"google-protobuf" library to create a "BinaryReader" instance, and at
some times, when the raw byte segment encodes a specific data format, it
also requires an low-level iterator to further decode each record.

This low-level iterator is only needed on certain circumstances but it's
being used to decode each and every single row. This makes the code more
complex and probably even slower.

Additionally, some data types expect the client to apply specific
padding rules. This is being done using custom routines for common
operations such as "leftPad" and "rightPad". The JavaScript API
available in the current minimum required version of Node.js (v14.0.0)
already includes native implementations of these functions, which can be
used instead.

This patch simplifies and optimizes the row parsing and decoding logic
by removing the need to use the low-level decoding iterator for most
records and replacing the custom "leftPad" and "rightPad" functions by
the corresponding native "padStart" and "padEnd" counterparts.

Change-Id: I2406276e088e8de448b7e6f58011008103136d86