BUG#34947621: fix Table.select() TypeScript definition
The Table.select() method accepts multiple optional arguments which can
be used to determine the columns that are included in the result set.

However, the TypeScript definition for this method does not specify the
possible types of these arguments leading to an error when TypeScript
application code is compiled to JavaScript using "tsc" (the TypeScript
compiler).

This patch updates the TypeScript definition to account for the possible
types of any argument provided to the Table.select() method, as
established by the X DevAPI specification.

https://dev.mysql.com/doc/x-devapi-userguide/en/crud-ebnf-table-crud-functions.html#crud-ebnf-tableselectfunction

Change-Id: Ie6322ef586aec8e28ffb5c86578dc72270357a85