BUG#34970001: move getAffectedItems() to SqlResult
The "RowResult" API is used to encapsulate the result set and
corresponding execution details of a data retrieval statement built
using the "Table.select()" API. This type of statements should not yield
any kind of side effects, including modifying data in the table, thus,
the outcome will not result in any affected items.

However, the "RowResult" API includes an extraneous "getAffectedItems()"
method, which is useless, given that it will always return 0. This
method is only relevant for the case where the statement is created
using raw SQL, in which case, the "Result" instance is not aware about
the type of statement that it yields from. A "SELECT" will not affect
any items, whereas an "UPDATE", "DELETE" or "INSERT" will.

This patch removes the "getAffectedItems()" from the "RowResult" API
whilst ensuring it is still available for results of executing a raw SQL
statement, via the corresponding "SqlResult" API.

Change-Id: I65d48a0cbde4a4452f1cc141e65fd80f985cac38