BUG#34970587: wrong getAutoIncrementValue() declaration
The "RowResult" module is used to create object instances containing the
result set, metadata and corresponding operational details yield when
executing a statement created with "Table.select()". The TypeScript
definition of its API includes a declaration for a
"getAutoIncrementValue()" method, which is not available in the native
JavaScript API.

This is because these kind of statements represent retrieval queries,
and not data modification operations. Executing such a statement should
not yield any new value for AUTO INCREMENT columns, thus leaving no need
for a method to obtain those values. The declaration is incorrect and
does not belong to that API.

This patch removes the extraneous "getAutoIncrementValue()" declaration
from the "RowResult" TypeScript API definition.

Change-Id: I6b1cebbe5e80326a6d3eb959d5eb7329b96cbbc8