WL#14254: Extended test suite with Docker environment
There is an extended set of tests distributed as part of the project
which are never executed within a regular test run. This is because
they require multiple MySQL server instances, or instances with a
specific configuration to be available.

Up until now, the testing infrastructure was optimized for a scenario
where it assumed that a MySQL server instance is available either in the
local system or on some other network-accessible host.

With the introduction of the Docker-based test setup in WL#15353, it is
now easier to address these limitations and migrate to a fully
container-based testing environment on top of something like Docker's
compose, where a testing container can communicate with other containers
that act as MySQL server instances and other service discovery
components enabling automated end-to-end tests for existing client
features such as Unix sockets, IPv6, TLS certificates, DNS SRV and
multi-host connections alongside effective regression and compatibility
testing with older MySQL versions.

This patch extends the existing Docker-based test setup to account for
all these scenarios, ensuring the full test suite can run in a
consistent and replicable setting.

Apart from the testing container, the setup now includes multiple
containers running MySQL server instances with different configurations.
Users are still allowed to run the default set of tests using an
instance available at a specific host. Otherwise, the tests run using
an instance on one of the containers, which is also used as a fallback
for tests using Unix sockets, on macOS, where it is not possible to
connect from the container to the host using that socket, or on Linux,
when the MYSQLX_SOCKET environment variable is not specified.

Change-Id: Ic7b4f524474ce527d047acff3ca8080bfbbc3f40