BUG#34959626: relax add() argument type strictness
With the existing TypeScript declaration for the "add()" method available
in both the "Collection" and "CollectionAdd" API definitions, it is not
possible to directly provide JavaScript objects obtained from a given
class instance. This is because the type definition for the method
arguments is too strict.

The reason for this is to prevent objects from containing fields whose
type is not capable of being encoded in the underlying protobuf message.
This behavior can also be relaxed wherein these fields can safely be
ignored, instead of preventing whole documents from being added to a
collection, resulting in a better user experience.

This patch relaxes the type strictness of the "add()" method arguments
and introduces lower level logic to ignore and skip values whose types
are unknown or cannot be encoded in X Protocol protobuf messages.

Change-Id: Ifdeacce0d9d11da4aaf4e997f7c9689031e750e2