tests: fix test failing with a parallel connection
A test related to prepared statements is failing when there are parallel
active connections to the database. This is because the test is querying
a performance_schema table to check if an index is being used with an
SQL statement that is not specific to the underlying thread id.

This patch updates the test to use a more specific SQL statement that
looks for the last instance of the given event that should be using the
index for the thread id associated to the uderlying connection used by
the test itself.

Change-Id: I810c03b799a8fc7682349c1501d1be2671ea688c