BUG#35666605: Allow expressions in "TableInsert.values()"
The "TableInsert.values()" API does not allow X DevAPI expressions to be
used for specifying table rows. There is no specific reason for this
functional limitation, when in fact, it is already possible to use X
DevAPI expressions to update a value in a colum.

Using expressions is useful to, for instance, define compatible GEOMETRY
column values in the client.

This patch introduces the changes required to allow inserting rows in a
table, when those rows are specified using an X DevAPI expression.

Change-Id: I16b027abd41f0aa2a8e718b0399db4bbebc08935