BUG#35707417: Return decimal value losing precision as string
Some values stored in DECIMAL and NUMERIC columns are losing precision
when returned back to the client after executing a table query. Although
some unsafe values are already returned as a raw string, others are not.
This is due to an issue in the decision-making process for when a
decimal value should be converted to JavaScript number or not.

This patch introduces the changes to address the shortcomings of that
decision-making process and ensures all unsafe values (integers or
decimals) are not converted to a JavaScript number.

Change-Id: Ica4016014a0f437b87e1cc797965eeba82ddfac1